Wire up TTS to new UI implementation. (#12)
Speech is wired to UI enough to debug/refactor. Still flaky.

We are on the road to having a more flexibly-designed app, but it's still unusable. The data layer needs to be hooked up, and the speech and UI layer are fragile and will likely crash frequently in addition to exhibiting other bugs.