Roughed in support for autoplay and autodelete.
I emphasize that the support is rough and buggy and doesn't work well yet. The entire codebase is due for a cleanup soon, and I am reluctant to over-optimize and hunt down timing bugs until then. Nevertheless, I'm merging into master so since master is not yet relied on by any party to be stable.